+++ lib/views/taxiDialog.dart
@@ -4,7 +4,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:open_store/open_store.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
-import 'package:taxiapp/constants/constants.dart';
 import 'package:taxiapp/constants/theme.dart';
 
 class TaxiDialog extends StatelessWidget {
